current status

+++ indexCardBackend/src/main/java/com/example/demo/Repositories/IndexCardRepository.java
@@ -6,5 +6,5 @@ import org.springframework.stereotype.Repository;
 import com.example.demo.entities.IndexCardEntity;
 
 @Repository
-public interface IndexCardRepository extends JpaRepository<IndexCardEntity, Long>{
+public interface IndexCardRepository extends JpaRepository<IndexCardEntity, Integer>{
 }
